+++ src/pyatmo/helpers.py
@@ -11,7 +11,7 @@ from .exceptions import NoDevice
 
 LOG: logging.Logger = logging.getLogger(__name__)
 
-_BASE_URL: str = "https://api.netatmo.com/"
+_DEFAULT_BASE_URL: str = "https://api.netatmo.com/"
 
 ERRORS: dict[int, str] = {
     400: "Bad request",
@@ -30,7 +30,7 @@ def to_time_string(value: str) -> str:
 
 
 def to_epoch(value: str) -> int:
-    return timegm(time.strptime(value + "GMT", "%Y-%m-%d_%H:%M:%S%Z"))
+    return timegm(time.strptime(f"{value}GMT", "%Y-%m-%d_%H:%M:%S%Z"))
 
 
 def today_stamps() -> tuple[int, int]: