+++ src/pyatmo/home_coach.py
@@ -1,8 +1,7 @@
 from .auth import AbstractAsyncAuth, NetatmoOAuth2
-from .helpers import _BASE_URL
 from .weather_station import AsyncWeatherStationData, WeatherStationData
 
-_GETHOMECOACHDATA_REQ = _BASE_URL + "api/gethomecoachsdata"
+_GETHOMECOACHDATA_ENDPOINT = "api/gethomecoachsdata"
 
 
 class HomeCoachData(WeatherStationData):
@@ -16,7 +15,7 @@ class HomeCoachData(WeatherStationData):
         Arguments:
             auth {NetatmoOAuth2} -- Authentication information with a valid access token
         """
-        super().__init__(auth, url_req=_GETHOMECOACHDATA_REQ, favorites=False)
+        super().__init__(auth, endpoint=_GETHOMECOACHDATA_ENDPOINT, favorites=False)
 
 
 class AsyncHomeCoachData(AsyncWeatherStationData):
@@ -30,4 +29,4 @@ class AsyncHomeCoachData(AsyncWeatherStationData):
         Arguments:
             auth {AbstractAsyncAuth} -- Authentication information with a valid access token
         """
-        super().__init__(auth, url_req=_GETHOMECOACHDATA_REQ, favorites=False)
+        super().__init__(auth, endpoint=_GETHOMECOACHDATA_ENDPOINT, favorites=False)