+++ tests/test_pyatmo_weatherstation.py
@@ -17,7 +17,8 @@ def test_weather_station_data(weather_station_data):
 
 def test_weather_station_data_no_response(auth, requests_mock):
     requests_mock.post(
-        pyatmo.weather_station._GETSTATIONDATA_REQ,
+        pyatmo.helpers._DEFAULT_BASE_URL
+        + pyatmo.weather_station._GETSTATIONDATA_ENDPOINT,
         json={},
         headers={"content-type": "application/json"},
     )
@@ -30,7 +31,8 @@ def test_weather_station_data_no_body(auth, requests_mock):
     with open("fixtures/status_ok.json", encoding="utf-8") as json_file:
         json_fixture = json.load(json_file)
     requests_mock.post(
-        pyatmo.weather_station._GETSTATIONDATA_REQ,
+        pyatmo.helpers._DEFAULT_BASE_URL
+        + pyatmo.weather_station._GETSTATIONDATA_ENDPOINT,
         json=json_fixture,
         headers={"content-type": "application/json"},
     )
@@ -43,7 +45,8 @@ def test_weather_station_data_no_data(auth, requests_mock):
     with open("fixtures/home_data_empty.json", encoding="utf-8") as json_file:
         json_fixture = json.load(json_file)
     requests_mock.post(
-        pyatmo.weather_station._GETSTATIONDATA_REQ,
+        pyatmo.helpers._DEFAULT_BASE_URL
+        + pyatmo.weather_station._GETSTATIONDATA_ENDPOINT,
         json=json_fixture,
         headers={"content-type": "application/json"},
     )
@@ -187,7 +190,7 @@ def test_weather_station_get_station(weather_station_data):
 def test_weather_station_get_module(weather_station_data, mid, expected):
     mod = weather_station_data.get_module(mid)
 
-    assert isinstance(mod, dict) is True
+    assert isinstance(mod, dict)
     assert mod.get("_id", mod) == expected
 
 
@@ -350,8 +353,7 @@ def test_weather_station_get_last_data(
     exclude,
     expected,
 ):
-    mod = weather_station_data.get_last_data(station_id, exclude=exclude)
-    if mod:
+    if mod := weather_station_data.get_last_data(station_id, exclude=exclude):
         assert sorted(mod) == expected
     else:
         assert mod == expected
@@ -417,8 +419,7 @@ def test_weather_station_check_updated(
     delay,
     expected,
 ):
-    mod = weather_station_data.check_updated(station_id, delay)
-    if mod:
+    if mod := weather_station_data.check_updated(station_id, delay):
         assert sorted(mod) == expected
     else:
         assert mod == expected
@@ -440,7 +441,7 @@ def test_weather_station_get_data(
     with open("fixtures/weatherstation_measure.json", encoding="utf-8") as json_file:
         json_fixture = json.load(json_file)
     requests_mock.post(
-        pyatmo.weather_station._GETMEASURE_REQ,
+        pyatmo.helpers._DEFAULT_BASE_URL + pyatmo.weather_station._GETMEASURE_ENDPOINT,
         json=json_fixture,
         headers={"content-type": "application/json"},
     )
@@ -492,8 +493,7 @@ def test_weather_station_get_last_data_bug_97(
     exclude,
     expected,
 ):
-    mod = weather_station_data.get_last_data(station_id, exclude)
-    if mod:
+    if mod := weather_station_data.get_last_data(station_id, exclude):
         assert sorted(mod) == expected
     else:
         assert mod == expected